Remove required restrictions on templates
Due to what appears to be a bug, when creating the template in
Visual Studio the required fields are never satisfied and stop the
project from being created. As this makes the templates worthless,
remove the required enforcement until the issue can be resolved.

The Authors and Description fields should still be considered required.