Remove pythonnet from nuspect
With Python template change to QuantConnect version of pythonnet,
a pythonnet folder is create. I was under the impression this was
required. It appear it is not copied transitively, it is hard to do so
with NuGet, an seems to only be necessary for backwards compatibility.
Lack of the folder blocks dotnet pack. Simple solution is to therefore
remove it. If this causes issues, more work will need to be done to
copy to output.